moving necessary files to the math module and removing fibonacci and inequality modules
