--- features/fibonacci/api/src/main/java/com/fredprojects/features/fibonacci/api/utils/CalculationStatus.kt
@@ -1,13 +1,0 @@
-package com.fredprojects.features.fibonacci.api.utils
-
-/**
- * The CalculationStatus is the status of the Fibonacci sequence calculation
- * @property SUCCESS if the calculation was successfully completed
- * @property INIT if the calculation is in progress
- * @property ERROR if the calculation failed
- */
-enum class CalculationStatus {
-    SUCCESS,
-    INIT,
-    ERROR
-}
