--- features/inequality/impl/.gitignore
@@ -1,1 +1,0 @@
-/build
