+++ features/math/src/main/java/com/fredprojects/features/math/domain/utils/CalculationStatus.kt
@@ -1,0 +1,7 @@
+package com.fredprojects.features.math.domain.utils
+
+enum class CalculationStatus {
+    SUCCESS,
+    INIT,
+    ERROR
+}
