+++ features/math/src/main/java/com/fredprojects/features/math/domain/utils/InequalityStatus.kt
@@ -1,0 +1,17 @@
+package com.fredprojects.features.math.domain.utils
+
+/**
+ * InequalityStatus represents the status of the inequality solution
+ * @property ERROR if an error has occurred during the InequalityUseCase execution or if the input is not valid
+ * @property NO_SOLUTION if there is no solution
+ * @property FIRST_SOLUTION if there is first type of solution
+ * @property SECOND_SOLUTION if there is second type of solution
+ * @property THIRD_SOLUTION if there is third type of solution
+ */
+enum class InequalityStatus {
+    ERROR,
+    NO_SOLUTION,
+    FIRST_SOLUTION,
+    SECOND_SOLUTION,
+    THIRD_SOLUTION
+}
